Remove two signed/unsigned comparison warnings
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@7317 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/space/common/basesetlocalkeystorage.hh
@@ -161,7 +161,7 @@ namespace Fem {
     //! access to stored entry with given geometry type 
     const Entry& operator [] ( const GeometryType& geomType ) const 
     {
-      assert( index( geomType ) < entryStorage_.size() );
+      assert( index( geomType ) < static_cast< int >( entryStorage_.size() ) );
       assert( entryStorage_[ index( geomType ) ] != 0 );
       return *( entryStorage_[ index( geomType ) ]);
     }
@@ -269,7 +269,7 @@ namespace Fem {
     inline const LocalKeyStorageType& compiledLocalKeys( const int order ) const 
     {
       assert( order - minPolOrder >= 0 );
-      assert( order - minPolOrder < int( compiledLocalKeys_.size() ) );
+      assert( int( order - minPolOrder ) < int( compiledLocalKeys_.size() ) );
       return compiledLocalKeys_[ order - minPolOrder ];
     }
